Completed framework
All the exists now would be quality of life changes - hash out the
playing field, make paddles look pretty, etc.  I may or may not return
to this at some point in the future.

If you are compiling this, you will need to link to the SFML 2.0 SDK,
and include DLLs.